moved app files into an app directory created a test directory where we can host the tests converted into a python module so that we can easily import stuff
